refactor(process): extract nextTick to a separate file
