+++ src/runtime/node/internal/process/nexttick.ts
@@ -1,33 +1,3 @@
-// https://nodejs.org/api/process.html#processhrtime
-export const hrtime: NodeJS.Process["hrtime"] = Object.assign(
-  function hrtime(startTime?: [number, number] | undefined) {
-    const now = Date.now();
-    // millis to seconds
-    const seconds = Math.trunc(now / 1000);
-    // convert millis to nanos
-    const nanos = (now % 1000) * 1_000_000;
-
-    if (startTime) {
-      let diffSeconds = seconds - startTime[0];
-      let diffNanos = nanos - startTime[0];
-
-      if (diffNanos < 0) {
-        diffSeconds = diffSeconds - 1;
-        diffNanos = 1_000_000_000 + diffNanos;
-      }
-      return [diffSeconds, diffNanos] as [number, number];
-    }
-
-    return [seconds, nanos] as [number, number];
-  },
-  {
-    bigint: function bigint() {
-      // Convert milliseconds to nanoseconds
-      return BigInt(Date.now() * 1_000_000);
-    },
-  },
-);
-
 // https://developer.mozilla.org/en-US/docs/Web/API/queueMicrotask
 // https://nodejs.org/api/process.html#when-to-use-queuemicrotask-vs-processnexttick
 export const nextTick: NodeJS.Process["nextTick"] = globalThis.queueMicrotask
